+++ dpppt-backend-sdk/dpppt-backend-sdk-ws/src/main/java/org/dpppt/backend/sdk/ws/radarcovid/insertmanager/insertionmodifier/SpanishEFGSModifier.java
@@ -44,6 +44,7 @@ public class SpanishEFGSModifier implements KeyInsertionModifier {
             gaenKey.setEfgsSharing(efgsSharing);
             gaenKey.setReportType(reportType);
             gaenKey.setDaysSinceOnsetOfSymptons(daysSinceOnsetOfSymtoms(onsetDay, gaenKey.getRollingStartNumber()));
+            gaenKey.getVisitedCountries().add(countryOrigin);
         }
 
         return content;